projects/redditflow/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Reddit
praw==7.7.1

# AI
anthropic==0.25.0
openai==1.12.0

# HTTP
httpx==0.26.0
requests==2.31.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring (optional)
sentry-sdk==1.39.2
